implement dagger android to details fragment, refactoring
