+++ app/src/main/java/com/intive/selftraining/selftraining/di/component/MovieApplicationComponent.kt
@@ -5,7 +5,8 @@ import com.intive.selftraining.selftraining.activity.di.binder.MovieActivityBind
 import com.intive.selftraining.selftraining.di.module.MovieApplicationModule
 import com.intive.selftraining.selftraining.di.module.NetworkModule
 import com.intive.selftraining.selftraining.di.module.UtilsModule
-import com.intive.selftraining.selftraining.listmovies.di.binder.MovieFragmentBinder
+import com.intive.selftraining.selftraining.listmovies.di.binder.MovieDetailsFragmentBinder
+import com.intive.selftraining.selftraining.listmovies.di.binder.ListMoviesFragmentBinder
 import dagger.Component
 import dagger.android.AndroidInjectionModule
 import javax.inject.Singleton
@@ -14,7 +15,7 @@ import javax.inject.Singleton
 @Component(
     modules = arrayOf(
         AndroidInjectionModule::class, MovieApplicationModule::class,
-        MovieActivityBinder::class, MovieFragmentBinder::class, NetworkModule::class, UtilsModule::class
+        MovieActivityBinder::class, ListMoviesFragmentBinder::class, MovieDetailsFragmentBinder::class, NetworkModule::class, UtilsModule::class
     )
 )
 interface MovieApplicationComponent {
