+++ app/src/main/java/com/intive/selftraining/selftraining/listmovies/di/factory/ListMoviesProviderFactory.kt
@@ -1,19 +1,19 @@
-package com.intive.selftraining.selftraining.di.providerFactory
+package com.intive.selftraining.selftraining.listmovies.di.factory
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.intive.selftraining.selftraining.movieDetails.MovieDetailsViewModel
-import com.intive.selftraining.selftraining.movieDetails.MovieRepository
+import com.intive.selftraining.selftraining.listmovies.ListMoviesRepository
+import com.intive.selftraining.selftraining.listmovies.ListMoviesViewModel
 import com.intive.selftraining.selftraining.network.CustomScheduler
 import com.intive.selftraining.selftraining.utils.ErrorHandler
 
-class MovieProviderFactory(
-    private val repo: MovieRepository,
+class ListMoviesProviderFactory(
+    private val repo: ListMoviesRepository,
     private val customScheduler: CustomScheduler,
     private val errorHandler: ErrorHandler
 ) : ViewModelProvider.Factory {
 
     override fun <T : ViewModel?> create(modelClass: Class<T>): T {
-        return MovieDetailsViewModel(repo, customScheduler, errorHandler) as T
+        return ListMoviesViewModel(repo, customScheduler, errorHandler) as T
     }
 }
