+++ app/src/main/java/com/intive/selftraining/selftraining/movieDetails/di/module/MovieDetailsRepoModule.kt
@@ -1,6 +1,6 @@
-package com.intive.selftraining.selftraining.di.module.movieDetails
+package com.intive.selftraining.selftraining.movieDetails.di.module
 
-import com.example.aleksandrtumanov.daggerattempts3.di.scopes.ActivityScope
+import com.intive.selftraining.selftraining.di.scopes.FragmentScope
 import com.intive.selftraining.selftraining.movieDetails.MovieRepository
 import com.intive.selftraining.selftraining.network.NetworkInterface
 import dagger.Module
@@ -9,7 +9,7 @@ import dagger.Provides
 @Module
 class MovieDetailsRepoModule {
 
-    @ActivityScope
+    @FragmentScope
     @Provides
     fun provideRepo(networkInterface: NetworkInterface): MovieRepository {
         return MovieRepository(networkInterface)
