+++ app/src/main/java/com/intive/selftraining/selftraining/movieDetails/di/module/MovieFragmentModule.kt
@@ -1,18 +1,20 @@
-package com.intive.selftraining.selftraining.di.module.movieDetails
+package com.intive.selftraining.selftraining.movieDetails.di.module
 
 import androidx.lifecycle.ViewModelProviders
-import com.intive.selftraining.selftraining.di.providerFactory.MovieProviderFactory
+import com.intive.selftraining.selftraining.di.scopes.FragmentScope
 import com.intive.selftraining.selftraining.movieDetails.MovieDetailsFragment
 import com.intive.selftraining.selftraining.movieDetails.MovieDetailsViewModel
 import com.intive.selftraining.selftraining.movieDetails.MovieRepository
+import com.intive.selftraining.selftraining.movieDetails.di.factory.MovieProviderFactory
 import com.intive.selftraining.selftraining.network.CustomScheduler
 import com.intive.selftraining.selftraining.utils.ErrorHandler
 import dagger.Module
 import dagger.Provides
 
 @Module
-class MovieFragmentModule(val fragment: MovieDetailsFragment) {
+class MovieFragmentModule() {
 
+    @FragmentScope
     @Provides
     fun providesMovieDetailsViewModelFactory(
         repo: MovieRepository,
@@ -26,8 +28,10 @@ class MovieFragmentModule(val fragment: MovieDetailsFragment) {
         )
     }
 
+    @FragmentScope
     @Provides
     fun providesMovieDetailsViewModel(
+        fragment: MovieDetailsFragment,
         factory: MovieProviderFactory
     ): MovieDetailsViewModel {
         return ViewModelProviders.of(fragment, factory).get(MovieDetailsViewModel::class.java)
